Add files via upload
version 1.2 : Added Routers. Multiple show screening. Final Ticket Template. Seat Booking final. Only those movies are shown which are up on screen.
